SE encontro el bug en el calculo de la diferencia impidiendo que este aumente del 50 porciento

+++ backend/src/main/java/com/api/payglobal/entity/Licencia.java
@@ -1,5 +1,6 @@
 package com.api.payglobal.entity;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 import com.fasterxml.jackson.annotation.JsonBackReference;
@@ -32,7 +33,7 @@ public class Licencia {
     private Integer limite;
     private Boolean activo;
     private LocalDate fechaCompra;
-    private Integer saldoAcumulado;
+    private BigDecimal saldoAcumulado;
 
     @OneToOne(mappedBy = "licencia")
     @JsonBackReference("licencia-mineria")
